+++ test/features/serialize-feature.js
@@ -1,7 +1,7 @@
 import testHelpers from '../helpers/testHelpers.js';
 import factory from '../helpers/factory.js';
 import * as Elements from 'bpmn-elements';
-import {default as Serializer, TypeResolver} from 'moddle-context-serializer';
+import {Serializer, TypeResolver} from 'moddle-context-serializer';
 import {extendFn} from '../../src/index.js';
 
 Feature('Extend function', () => {
@@ -53,6 +53,38 @@ Feature('Extend function', () => {
       });
       expect(registered.script).to.have.property('scriptFormat');
       expect(registered.script).to.have.property('body');
+    });
+
+    Given('a source with sequence flow execution listener', async () => {
+      const source = factory.resource('sequence-flow-properties.bpmn');
+      moddleContext = await testHelpers.moddleContext(source, await testHelpers.getModdleExtensions());
+    });
+
+    When('serialized', () => {
+      serialized = Serializer(moddleContext, TypeResolver(Elements), extendFn);
+    });
+
+    Then('all scripts are registered', () => {
+      expect(serialized.elements.scripts).to.have.length(4);
+    });
+
+    And('registered SequenceFlow execution listener scripts', () => {
+      const [listener0, listener1] = serialized.elements.scripts.filter((s) => s.parent.id === 'to-script');
+      expect(listener0).to.have.property('name', 'to-script/camunda:ExecutionListener/take/0');
+      expect(listener0).to.have.property('parent').that.deep.equal({
+        id: 'to-script',
+        type: 'bpmn:SequenceFlow',
+      });
+      expect(listener0.script).to.have.property('scriptFormat');
+      expect(listener0.script).to.have.property('body');
+
+      expect(listener1).to.have.property('name', 'to-script/camunda:ExecutionListener/take/1');
+      expect(listener1).to.have.property('parent').that.deep.equal({
+        id: 'to-script',
+        type: 'bpmn:SequenceFlow',
+      });
+      expect(listener1.script).to.have.property('scriptFormat');
+      expect(listener1.script).to.have.property('resource');
     });
   });
 });